# feat: Opt for concatenation instead of using fmt.Sprintf